* NEWS: Update with latest news. * magick/blob.c (WriteBlob): Move BlobStream write support to a subroutine for easier maintenance. * coders/dpx.c (ReadDPXImage): Support retrieving all DPX attributes as image attributes. (WriteDPXImage): Buffer writes for better performance on some platforms.
* coders/dpx.c (WriteDPXImage): Re-write sample marshalling to be
based on a series of tighter loops.  Results in a small
performance increase.
(ReadDPXImage): Re-write sample marshalling to be
based on a series of tighter loops.  Results in a small
performance increase.

+++ coders/fits.c
@@ -735,7 +735,7 @@ static unsigned int WriteFITSImage(const ImageInfo *image_info,Image *image)
   (void) strncpy(fits_info+320,buffer,strlen(buffer));
   FormatString(buffer,"DATAMIN =           %10u",0);
   (void) strncpy(fits_info+400,buffer,strlen(buffer));
-  FormatString(buffer,"DATAMAX =           %10u",MaxValueGivenBits(depth));
+  FormatString(buffer,"DATAMAX =           %10lu",(unsigned long) MaxValueGivenBits(depth));
   (void) strncpy(fits_info+480,buffer,strlen(buffer));
   (void) strcpy(buffer,"HISTORY Created by GraphicsMagick.");
   (void) strncpy(fits_info+560,buffer,strlen(buffer));
